+++ MatchThree.Domain/Interfaces/LeaderboardMember/IReadLeaderboardMemberService.cs
@@ -6,5 +6,5 @@ namespace MatchThree.Domain.Interfaces.LeaderboardMember;
 public interface IReadLeaderboardMemberService
 {
     Task<IReadOnlyCollection<LeaderboardMemberEntity>> GetLeaderboardByLeagueAsync(LeagueTypes league);
-    Task<ushort> GetTopSpotByUserId(long userId);
+    Task<int> GetTopSpotByUserId(long userId);
 }
