+++ MatchThree.Domain/Models/LeaderboardMemberEntity.cs
@@ -6,7 +6,7 @@ public class LeaderboardMemberEntity
 {
     public long Id { get; set; }
     public LeagueTypes League { get; set; }
-    public ushort TopSpot { get; set; }
+    public int TopSpot { get; set; }
     public string FirstName { get; set; } = string.Empty;
     public ulong OverallBalance { get; set; }
 }
