+++ MatchThree/Models/BalanceDto.cs
@@ -7,5 +7,5 @@ public class BalanceDto
     public long Id { get; set; }
     public uint Balance { get; set; }
     public LeagueTypes League { get; set; }
-    public ushort TopSpot { get; set; }
+    public int TopSpot { get; set; }
 }
